+++ src/main/java/com/google/firebase/auth/ErrorInfo.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2018 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.firebase.auth;
+
+public class ErrorInfo {
+
+  private final int index;
+  private final String reason;
+
+  ErrorInfo(int index, String reason) {
+    this.index = index;
+    this.reason = reason;
+  }
+
+  public int getIndex() {
+    return index;
+  }
+
+  public String getReason() {
+    return reason;
+  }
+}